feat: US-001 - Put shared display helpers in the right layer
