+++ docs/processes/development-guide-relevant-files.md
@@ -74,6 +74,8 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/App.test.tsx` | Dashboard app-level regression surface | Mixed selected-work dispatch history coverage belongs here when a change needs full-app proof that inference-backed and script-backed workstation requests still share one `Workstation dispatches` surface. |
 | `ui/src/components/ui/agent-bento.stories.tsx` | Shared bento-layout Storybook seam | Shared layout cleanup work should keep at least one real multi-card dashboard state here, using actual feature cards rather than placeholder children, so ownership moves still prove board rendering and widget placement through the relocated primitive seam. |
 | `ui/src/components/ui/widget-frame.stories.tsx` | Shared widget-frame Storybook seam | Widget-frame ownership changes should keep a real feature framing state here, such as current-selection empty guidance, so Storybook evidence continues to cover observable header controls, card titles, and empty/detail body treatment through the shared frame owner. |
+| `ui/src/components/ui/formatters.ts` | Shared display formatter seam | Cross-feature duration, work-item, provider-session, and trace display helpers belong here; any legacy `components/dashboard/formatters.ts` path should stay a thin compatibility re-export instead of regaining primary behavior ownership. |
+| `ui/src/components/ui/place-labels.ts` | Shared place-label seam | Cross-feature place and place-state label helpers belong here; compatibility `components/dashboard/place-labels.ts` imports should stay shims rather than becoming a second implementation owner. |
 | `ui/integration/event-stream-replay.integration.test.mjs` | Browser-backed dashboard integration harness | Keep browser-visible dashboard smoke, adjacent API stubs such as `/factory/~current` and cross-origin `POST /factory`, drag/drop `DataTransfer` file setup, and blob-download artifact capture in this one preview-backed Playwright/Vitest lane so UI integration coverage extends the built app without adding a second browser harness. |
 | `ui/src/components/ui/` | Shared website primitive layer | New reusable button, field, dialog, chart, table, skeleton, calendar, collapsible, resizable, and shared layout primitives such as bento boards or widget frames belong here; when legacy dashboard import paths still need compatibility, keep `components/dashboard/*` as thin re-export shims instead of the primary implementation owner. |
 | `ui/src/hooks/dashboard/useDashboardLayout.ts` | Dashboard layout persistence seam | The persisted layout contract should import `AgentBentoLayoutItem` from `ui/src/components/ui/` so local-storage migrations, default card positions, and the live bento grid stay aligned with the shared layout owner instead of a compatibility shim. |
