+++ ui/src/components/ui/index.ts
@@ -6,8 +6,10 @@ export * from "./collapsible";
 export * from "./data-table";
 export * from "./dashboard-typography";
 export * from "./dialog";
+export * from "./formatters";
 export * from "./input";
 export * from "./classnames";
+export * from "./place-labels";
 export * from "./resizable";
 export * from "./select";
 export * from "./skeleton";
